Add builder usage docs
